chore: upgrade TypeScript to 4.7
