+++ package.json
@@ -40,28 +40,28 @@
     "patches"
   ],
   "devDependencies": {
-    "@auto-it/conventional-commits": "^10.32.6",
-    "@auto-it/first-time-contributor": "^10.32.6",
-    "@babel/plugin-transform-runtime": "^7.17.0",
+    "@auto-it/conventional-commits": "^10.37.1",
+    "@auto-it/first-time-contributor": "^10.37.1",
+    "@babel/plugin-transform-runtime": "^7.18.2",
     "@rollup/plugin-commonjs": "22.0.0",
     "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^13.1.3",
+    "@rollup/plugin-node-resolve": "^13.3.0",
     "@types/semver": "^7.3.9",
-    "auto": "^10.32.6",
-    "browser-sync": "^2.27.7",
-    "concurrently": "^7.0.0",
+    "auto": "^10.37.1",
+    "browser-sync": "^2.27.10",
+    "concurrently": "^7.2.1",
     "del-cli": "^4.0.1",
-    "jest-canvas-mock": "^2.3.1",
+    "jest-canvas-mock": "^2.4.0",
     "patch-package": "^6.4.7",
     "postinstall-postinstall": "^2.1.0",
-    "rollup": "2.75.3",
+    "rollup": "2.75.4",
     "rollup-plugin-bundle-size": "^1.0.3",
     "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-ts": "^2.0.5",
-    "sass": "^1.49.9",
-    "typescript": "^4.6.2",
-    "vega": "^5.21.0",
-    "vega-lite": "^5.0.0",
+    "rollup-plugin-ts": "^3.0.0",
+    "sass": "^1.52.1",
+    "typescript": "^4.7.2",
+    "vega": "^5.22.1",
+    "vega-lite": "^5.2.0",
     "vega-lite-dev-config": "^0.20.0"
   },
   "peerDependencies": {
@@ -69,10 +69,10 @@
     "vega-lite": "*"
   },
   "dependencies": {
-    "fast-json-patch": "^3.1.0",
+    "fast-json-patch": "^3.1.1",
     "json-stringify-pretty-compact": "^3.0.0",
-    "semver": "^7.3.5",
-    "tslib": "^2.3.1",
+    "semver": "^7.3.7",
+    "tslib": "^2.4.0",
     "vega-interpreter": "^1.0.4",
     "vega-schema-url-parser": "^2.2.0",
     "vega-themes": "^2.10.0",
